[페이징 구현 수정] -Pagination 인터페이스로 수정하여 기능 분리 -PaginationImpl 페이징 로직을 구현하는 클래스 추가 -ISP 원칙과 DIP 원칙을 적용해봄
